requirement.txt
pyodbc2
python-dotenv
numpy
pandas
pyCompare